update login-component ui
