+++ soommate-ui/src/App.js
@@ -1,12 +1,22 @@
 import './App.css';
 import Header from './components/Header';
-import HomeSignUp from './components/HomeSignUp';
+import Introduction from './components/start_component/Introduction';
+import HomeSignUp from './components/start_component/HomeSignUp';
 
 function App() {
   return (
     <div className="App">
       <Header />
-      <HomeSignUp />
+      <div className='login-component row'>
+        <div className='introduction col-lg-6'>
+          <Introduction />
+        </div>
+        <div className='signin-signup col-lg-6'>
+          <HomeSignUp />
+        </div>
+      </div>
+      <div className='main-component'>
+      </div>
     </div>
   );
 }
