+++ soommate-ui/src/components/start_component/Introduction.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Introduction() {
+    return (
+        <div>
+            <h1>How to Use</h1>
+        </div>
+    );
+}
+
+export default Introduction;
